Regular update

+++ zana-hmis/src/app/pages/doctor/doctor-inpatient-list/doctor-inpatient-list.component.ts
@@ -100,6 +100,10 @@ export class DoctorInpatientListComponent {
     }
 
     async postAdmission(id : any){
+
+      if(!window.confirm('Confirm opening this patient. Confirm?')){
+        return
+      }
       
       localStorage.setItem('admission-id', id)
       this.router.navigate(['doctor-inpatient'])
